Move Watchers component alongside jsdebugger components where its used
