+++ apps/src/lib/tools/jsdebugger/Watchers.jsx
@@ -5,7 +5,7 @@ import {connect} from 'react-redux';
 import TetherComponent from 'react-tether';
 
 import i18n from '@cdo/locale';
-import {add, update, remove} from '../../redux/watchedExpressions';
+import {add, update, remove} from '../../../redux/watchedExpressions';
 import AutocompleteSelector from './AutocompleteSelector';
 
 import escapeSpecialCharactersForRegExp from '@cdo/apps/util/escapeSpecialCharactersForRegExp';
