+++ apps/test/unit/lib/tools/jsdebugger/AutocompleteSelectorTest.jsx
@@ -1,9 +1,9 @@
-import {expect} from '../../../util/deprecatedChai';
+import {expect} from '../../../../util/deprecatedChai';
 import sinon from 'sinon';
 import React from 'react';
 import {mount} from 'enzyme';
-import AutocompleteSelector from '@cdo/apps/templates/watchers/AutocompleteSelector';
-import {allowConsoleWarnings} from '../../../util/testUtils';
+import AutocompleteSelector from '@cdo/apps/lib/tools/jsdebugger/AutocompleteSelector';
+import {allowConsoleWarnings} from '../../../../util/testUtils';
 
 describe('AutocompleteSelector', () => {
   // TODO: (madelynkasula) Silences componentWillReceiveProps deprecation warning due to React 16 upgrade.
